Entrega lista

+++ src/routes/products.router.js
@@ -54,8 +54,12 @@ router.put("/:pid", async (req, res) => {
         const id = req.params.pid;
         const data = req.body;
 
+        const product = await productHandling.getProductById(id);
+
+        if (!product) return res.status(404).send({ status: "error", message: "Product not found" });
+
         if (!req.session.user) return res.status(401).send(errors.login);
-        if (req.session.user.level < 1) return res.status(401).send(errors.lowPerms);
+        if (req.session.user.level < 1 || product.owner != req.session.user._id) return res.status(401).send(errors.lowPerms);
 
         productHandling.updateProduct(id, data);
         res.send({ status: "success" });
@@ -72,8 +76,12 @@ router.delete("/:pid", async (req, res) => {
     try {
         const id = req.params.pid;
 
+        const product = await productHandling.getProductById(id);
+
+        if (!product) return res.status(404).send({ status: "error", message: "Product not found" });
+
         if (!req.session.user) return res.status(401).send(errors.login);
-        if (req.session.user.level < 1) return res.status(401).send(errors.lowPerms);
+        if (req.session.user.level < 1 || product.owner != req.session.user._id) return res.status(401).send(errors.lowPerms);
 
         productHandling.deleteProduct(id);
         res.send({ status: "success" });
@@ -91,7 +99,9 @@ router.post("/", async (req, res) => {
         const product = req.body;
 
         if (!req.session.user) return res.status(401).send(errors.login);
-        if (req.session.user.level < 1) return res.status(401).send(errors.lowPerms);
+        if (req.session.user.level < 1 || req.session.user.premium != true) return res.status(401).send(errors.lowPerms);
+
+        product.owner = req.session.user._id;
 
         productHandling.addProduct(product);
         res.send({ status: "success" });
